ディレクトリ構成を変更 - 完全に好みだがpagesやcomponentsをsrc配下に設定 - こうすることでソースコードであると明示的に示すとともに設定のためのファイルたちと一線を画すと考えています
