+++ src/pages/index.tsx
@@ -1,0 +1,23 @@
+import type { NextPage } from 'next';
+import Head from 'next/head';
+
+import { ReactElement, useState } from 'react';
+
+import { Layout } from 'src/components/Layout';
+import { About } from 'src/components/About';
+
+const Home: NextPage = () => {
+  const [children, setChildren] = useState<ReactElement>(<About />);
+
+  return (
+    <>
+      <Head>
+        <title>SecBokApp</title>
+      </Head>
+
+      <Layout setChildren={setChildren}>{children}</Layout>
+    </>
+  );
+};
+
+export default Home;
